Move generic request setup logic to setup-request.js
Pass `createRequestFn` as parameter to the generic `makeRequest` function.

Set scopes to `accounts payments` for `client_credentials` grant.

Intention is this code be reused for payments request.
