+++ app/setup-request/index.js
@@ -1,0 +1,3 @@
+const { makeRequest } = require('./setup-request');
+
+exports.makeRequest = makeRequest;
